restructure code into 3 subpackages for backend frontend and shared
